+++ src/textual/messages.py
@@ -9,7 +9,6 @@ from .geometry import Region
 from .message import Message
 
 if TYPE_CHECKING:
-    from .message_pump import MessagePump
     from .widget import Widget
 
 
@@ -25,12 +24,11 @@ class ExitApp(Message, verbose=True):
 
 @rich.repr.auto
 class Update(Message, verbose=True):
-    def __init__(self, sender: MessagePump, widget: Widget):
-        super().__init__(sender)
+    def __init__(self, widget: Widget):
+        super().__init__()
         self.widget = widget
 
     def __rich_repr__(self) -> rich.repr.Result:
-        yield self.sender
         yield self.widget
 
     def __eq__(self, other: object) -> bool:
@@ -63,9 +61,9 @@ class UpdateScroll(Message, verbose=True):
 class InvokeLater(Message, verbose=True, bubble=False):
     """Sent by Textual to invoke a callback."""
 
-    def __init__(self, sender: MessagePump, callback: CallbackType) -> None:
+    def __init__(self, callback: CallbackType) -> None:
         self.callback = callback
-        super().__init__(sender)
+        super().__init__()
 
     def __rich_repr__(self) -> rich.repr.Result:
         yield "callback", self.callback
@@ -75,9 +73,9 @@ class InvokeLater(Message, verbose=True, bubble=False):
 class ScrollToRegion(Message, bubble=False):
     """Ask the parent to scroll a given region in to view."""
 
-    def __init__(self, sender: MessagePump, region: Region) -> None:
+    def __init__(self, region: Region) -> None:
         self.region = region
-        super().__init__(sender)
+        super().__init__()
 
 
 class Prompt(Message, no_dispatch=True):