+++ src/textual/widgets/_button.py
@@ -165,9 +165,14 @@ class Button(Static, can_focus=True):
             button: The button that was pressed.
         """
 
+        def __init__(self, button: Button) -> None:
+            self.button = button
+            super().__init__()
+
         @property
-        def button(self) -> Button:
-            return cast(Button, self.sender)
+        def control(self) -> Button:
+            """Alias for the button."""
+            return self.button
 
     def __init__(
         self,
@@ -235,7 +240,7 @@ class Button(Static, can_focus=True):
         # Manage the "active" effect:
         self._start_active_affect()
         # ...and let other components know that we've just been clicked:
-        self.post_message_no_wait(Button.Pressed(self))
+        self.post_message(Button.Pressed(self))
 
     def _start_active_affect(self) -> None:
         """Start a small animation to show the button was clicked."""
@@ -247,7 +252,7 @@ class Button(Static, can_focus=True):
     async def _on_key(self, event: events.Key) -> None:
         if event.key == "enter" and not self.disabled:
             self._start_active_affect()
-            await self.post_message(Button.Pressed(self))
+            self.post_message(Button.Pressed(self))
 
     @classmethod
     def success(