+++ src/textual/widgets/_list_item.py
@@ -1,5 +1,7 @@
 """Provides a list item widget for use with `ListView`."""
 
+from __future__ import annotations
+
 from textual import events
 from textual.message import Message
 from textual.reactive import reactive
@@ -41,10 +43,12 @@ class ListItem(Widget, can_focus=False):
     class _ChildClicked(Message):
         """For informing with the parent ListView that we were clicked"""
 
-        sender: "ListItem"
+        def __init__(self, item: ListItem) -> None:
+            self.item = item
+            super().__init__()
 
     def on_click(self, event: events.Click) -> None:
-        self.post_message_no_wait(self._ChildClicked(self))
+        self.post_message(self._ChildClicked(self))
 
     def watch_highlighted(self, value: bool) -> None:
         self.set_class(value, "--highlight")