+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.14.0] - Unreleased
+
+### Changes
+
+- Breaking change: There is now only `post_message` to post events, which is non-async, `post_message_no_wait` was dropped. https://github.com/Textualize/textual/pull/1940
+- Breaking change: The Timer class now has just one method to stop it, `Timer.stop` which is non sync https://github.com/Textualize/textual/pull/1940
+- Breaking change: Messages don't require a `sender` in their constructor https://github.com/Textualize/textual/pull/1940
+- Many messages have grown a `control` property which returns the control they relate to. https://github.com/Textualize/textual/pull/1940
+- Dropped `time` attribute from Messages https://github.com/Textualize/textual/pull/1940
+
+### Added
+
+- Added `data_table` attribute to DataTable events https://github.com/Textualize/textual/pull/1940
+- Added `list_view` attribute to `ListView` events https://github.com/Textualize/textual/pull/1940
+- Added `radio_set` attribute to `RadioSet` events https://github.com/Textualize/textual/pull/1940
+- Added `switch` attribute to `Switch` events https://github.com/Textualize/textual/pull/1940
+- Breaking change: Added `toggle_button` attribute to RadioButton and Checkbox events, replaces `input` https://github.com/Textualize/textual/pull/1940
+
 ## [0.13.0] - 2023-03-02
 
 ### Added